nbplkup: add -s option to produce more script-friendly output
If the -s option is specified, no attempt will be made to prettify the
results into a table; instead, the address and socket number will be
printed first, followed by a single space, followed by the NBP name.
This should allow this output to be unambiguously parsed by scripts
using regexes.